+++ lib/src/inspector_controller.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:shake/shake.dart';
+import 'package:requests_inspector/src/shake.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../requests_inspector.dart';